fix: api 모듈 전역기능 패키기명 global로 수정 [ISSUE-#7]
