+++ application/api/src/main/java/com/vrr/ApiApplication.java
@@ -1,6 +1,6 @@
 package com.vrr;
 
-import com.vrr.libs.auth.properties.AppProperties;
+import com.vrr.global.auth.properties.AppProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
