+++ application/api/src/main/java/com/vrr/global/auth/exception/OAuthProviderMissMatchException.java
@@ -1,0 +1,7 @@
+package com.vrr.global.auth.exception;
+
+public class OAuthProviderMissMatchException extends RuntimeException {
+    public OAuthProviderMissMatchException(String message) {
+        super(message);
+    }
+}
