+++ application/api/src/main/java/com/vrr/global/auth/exception/TokenNotValidException.java
@@ -1,0 +1,12 @@
+package com.vrr.global.auth.exception;
+
+public class TokenNotValidException extends RuntimeException{
+
+    public TokenNotValidException() {
+        super("Failed to generate Token.");
+    }
+
+    private TokenNotValidException(String message) {
+        super(message);
+    }
+}
