+++ application/api/src/main/java/com/vrr/global/auth/token/HeaderTokenParser.java
@@ -1,0 +1,23 @@
+package com.vrr.global.auth.token;
+
+import javax.servlet.http.HttpServletRequest;
+
+public class HeaderTokenParser {
+
+    private final static String HEADER_AUTHORIZATION = "Authorization";
+    private final static String TOKEN_PREFIX = "Bearer ";
+
+    public static String getAccessToken(HttpServletRequest request) {
+        String headerValue = request.getHeader(HEADER_AUTHORIZATION);
+
+        if (headerValue == null) {
+            return null;
+        }
+
+        if (headerValue.startsWith(TOKEN_PREFIX)) {
+            return headerValue.substring(TOKEN_PREFIX.length());
+        }
+
+        return null;
+    }
+}
