+++ application/api/src/main/java/com/vrr/global/config/JwtConfig.java
@@ -1,0 +1,18 @@
+package com.vrr.global.config;
+
+import com.vrr.global.auth.token.AuthTokenProvider;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class JwtConfig {
+
+    @Value("${jwt.secret}")
+    private String secret;
+
+    @Bean
+    public AuthTokenProvider jwtProvider() {
+        return new AuthTokenProvider(secret);
+    }
+}
