+++ application/api/src/main/java/com/vrr/global/config/SecurityBean.java
@@ -1,0 +1,26 @@
+package com.vrr.global.config;
+
+import com.vrr.global.auth.filter.TokenAuthenticationFilter;
+import com.vrr.global.auth.token.AuthTokenProvider;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+@RequiredArgsConstructor
+public class SecurityBean {
+
+    private final AuthTokenProvider tokenProvider;
+
+    @Bean
+    public TokenAuthenticationFilter tokenAuthenticationFilter() {
+        return new TokenAuthenticationFilter(tokenProvider);
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+}
