--- application/api/src/main/java/com/vrr/libs/auth/exception/OAuthProviderMissMatchException.java
@@ -1,7 +1,0 @@
-package com.vrr.libs.auth.exception;
-
-public class OAuthProviderMissMatchException extends RuntimeException {
-    public OAuthProviderMissMatchException(String message) {
-        super(message);
-    }
-}
