--- application/api/src/main/java/com/vrr/libs/auth/exception/TokenNotValidException.java
@@ -1,12 +1,0 @@
-package com.vrr.libs.auth.exception;
-
-public class TokenNotValidException extends RuntimeException{
-
-    public TokenNotValidException() {
-        super("Failed to generate Token.");
-    }
-
-    private TokenNotValidException(String message) {
-        super(message);
-    }
-}
