--- application/api/src/main/java/com/vrr/libs/auth/handler/OAuth2AuthenticationFailerHandler.java
@@ -1,41 +1,0 @@
-package com.vrr.libs.auth.handler;
-
-import com.vrr.libs.auth.repository.HttpCookieOAuth2AuthorizationRequestRepository;
-import com.vrr.libs.utils.CookieUtils;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
-import org.springframework.web.util.UriComponentsBuilder;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-import static com.vrr.libs.auth.repository.HttpCookieOAuth2AuthorizationRequestRepository.REDIRECT_URI_PARAM_COOKIE_NAME;
-
-@Component
-@RequiredArgsConstructor
-public class OAuth2AuthenticationFailerHandler extends SimpleUrlAuthenticationFailureHandler {
-
-    private final HttpCookieOAuth2AuthorizationRequestRepository authorizationRequestRepository;
-
-    @Override
-    public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException, ServletException {
-        String targetUrl = CookieUtils.getCookie(request, REDIRECT_URI_PARAM_COOKIE_NAME)
-                .map(Cookie::getValue)
-                .orElse(("/"));
-
-        exception.printStackTrace();
-
-        targetUrl = UriComponentsBuilder.fromUriString(targetUrl)
-                .queryParam("error", exception.getLocalizedMessage())
-                .build().toUriString();
-
-        authorizationRequestRepository.removeAuthorizationRequestCookies(request, response);
-
-        getRedirectStrategy().sendRedirect(request, response, targetUrl);
-    }
-}
