--- application/api/src/main/java/com/vrr/libs/auth/info/OAuth2UserInfo.java
@@ -1,24 +1,0 @@
-package com.vrr.libs.auth.info;
-
-import java.util.Map;
-
-public abstract class OAuth2UserInfo {
-
-    protected Map<String, Object> attributes;
-
-    public OAuth2UserInfo(Map<String, Object> attributes) {
-        this.attributes = attributes;
-    }
-
-    public Map<String, Object> getAttributes() {
-        return attributes;
-    }
-
-    public abstract String getId();
-
-    public abstract String getName();
-
-    public abstract String getEmail();
-
-    public abstract String getImageUrl();
-}
