--- application/api/src/main/java/com/vrr/libs/auth/info/implement/GoogleOAuth2UserInfo.java
@@ -1,32 +1,0 @@
-package com.vrr.libs.auth.info.implement;
-
-import com.vrr.libs.auth.info.OAuth2UserInfo;
-
-import java.util.Map;
-
-public class GoogleOAuth2UserInfo extends OAuth2UserInfo {
-
-    public GoogleOAuth2UserInfo(Map<String, Object> attributes) {
-        super(attributes);
-    }
-
-    @Override
-    public String getId() {
-        return (String) attributes.get("sub");
-    }
-
-    @Override
-    public String getName() {
-        return (String) attributes.get("name");
-    }
-
-    @Override
-    public String getEmail() {
-        return (String) attributes.get("email");
-    }
-
-    @Override
-    public String getImageUrl() {
-        return (String) attributes.get("picture");
-    }
-}
