--- application/api/src/main/java/com/vrr/libs/auth/info/implement/NaverOAuth2UserInfo.java
@@ -1,56 +1,0 @@
-package com.vrr.libs.auth.info.implement;
-
-import com.vrr.libs.auth.info.OAuth2UserInfo;
-
-import java.util.Map;
-
-public class NaverOAuth2UserInfo extends OAuth2UserInfo {
-
-    public NaverOAuth2UserInfo(Map<String, Object> attributes) {
-        super(attributes);
-    }
-
-    @Override
-    public String getId() {
-        Map<String, Object> response = (Map<String, Object>) attributes.get("response");
-
-        if (response == null) {
-            return null;
-        }
-
-        return (String) response.get("id");
-    }
-
-    @Override
-    public String getName() {
-        Map<String, Object> response = (Map<String, Object>) attributes.get("response");
-
-        if (response == null) {
-            return null;
-        }
-
-        return (String) response.get("nickname");
-    }
-
-    @Override
-    public String getEmail() {
-        Map<String, Object> response = (Map<String, Object>) attributes.get("response");
-
-        if (response == null) {
-            return null;
-        }
-
-        return (String) response.get("email");
-    }
-
-    @Override
-    public String getImageUrl() {
-        Map<String, Object> response = (Map<String, Object>) attributes.get("response");
-
-        if (response == null) {
-            return null;
-        }
-
-        return (String) response.get("profile_image");
-    }
-}
