--- application/api/src/main/java/com/vrr/libs/auth/properties/AppProperties.java
@@ -1,41 +1,0 @@
-package com.vrr.libs.auth.properties;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Getter
-@ConfigurationProperties(prefix = "app")
-public class AppProperties {
-
-    private final Auth auth = new Auth();
-    private final OAuth2 oauth2 = new OAuth2();
-
-    @Getter
-    @Setter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class Auth {
-        private String tokenSecret;
-        private long tokenExpiry;
-        private long refreshTokenExpiry;
-    }
-
-    public static final class OAuth2 {
-        private List<String> authorizedRedirectUris = new ArrayList<>();
-
-        public List<String> getAuthorizedRedirectUris() {
-            return authorizedRedirectUris;
-        }
-
-        public OAuth2 authorizedRedirectUris(List<String> authorizedRedirectUris) {
-            this.authorizedRedirectUris = authorizedRedirectUris;
-            return this;
-        }
-    }
-}
