--- application/api/src/main/java/com/vrr/libs/auth/service/OAuth2UserServiceImpl.java
@@ -1,95 +1,0 @@
-package com.vrr.libs.auth.service;
-
-import com.vrr.code.auth.ProviderType;
-import com.vrr.code.auth.RoleType;
-import com.vrr.entity.auth.User;
-import com.vrr.entity.auth.UserRepository;
-import com.vrr.libs.auth.exception.OAuthProviderMissMatchException;
-import com.vrr.libs.auth.info.OAuth2UserInfo;
-import com.vrr.libs.auth.info.OAuth2UserInfoFactory;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.InternalAuthenticationServiceException;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.stereotype.Service;
-
-import javax.naming.AuthenticationException;
-import java.time.LocalDateTime;
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-public class OAuth2UserServiceImpl extends DefaultOAuth2UserService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
-        OAuth2User user = super.loadUser(userRequest);
-
-        try {
-            return this.process(userRequest, user);
-        }  catch (Exception ex) {
-            if (ex instanceof AuthenticationException) {
-                throw ex;
-            }
-
-            ex.printStackTrace();
-            throw new InternalAuthenticationServiceException(ex.getMessage(), ex.getCause());
-        }
-    }
-
-    private OAuth2User process(OAuth2UserRequest userRequest, OAuth2User user) {
-        ProviderType providerType = ProviderType.valueOf(userRequest.getClientRegistration().getRegistrationId().toUpperCase());
-
-        OAuth2UserInfo userInfo = OAuth2UserInfoFactory.getOAuth2UserInfo(providerType, user.getAttributes());
-        Optional<User> userOptional = userRepository.findBySerialNumber(userInfo.getId());
-
-        User savedUser;
-        if (userOptional.isPresent()) {
-            savedUser = userOptional.get();
-            if (providerType != savedUser.getProviderType()) {
-                throw new OAuthProviderMissMatchException(
-                        "Looks like you're signed up with " + providerType + "account. " +
-                        "Please use your " + savedUser.getProviderType() + " account to login."
-                );
-            }
-            savedUser = updateUser(savedUser, userInfo);
-        } else {
-            savedUser = createUser(userInfo, providerType);
-        }
-
-        return UserPrincipal.create(savedUser, user.getAttributes());
-    }
-
-    private User createUser(OAuth2UserInfo userInfo, ProviderType providerType) {
-        LocalDateTime now = LocalDateTime.now();
-        User user = new User(
-                userInfo.getId(),
-                userInfo.getName(),
-                userInfo.getEmail(),
-                "Y",
-                userInfo.getImageUrl(),
-                providerType,
-                RoleType.USER,
-                now,
-                now
-        );
-
-        return userRepository.save(user);
-    }
-
-    private User updateUser(User user, OAuth2UserInfo userInfo) {
-        if (userInfo.getName() != null && !user.getUsername().equals(userInfo.getName())) {
-            user.updateProfile(userInfo.getName(), null);
-        }
-
-        if (userInfo.getImageUrl() != null && !user.getProfileImageUrl().equals(userInfo.getImageUrl())) {
-            user.updateProfile(null, userInfo.getImageUrl());
-        }
-
-        return userRepository.save(user);
-    }
-}
