--- application/api/src/main/java/com/vrr/libs/auth/token/AuthTokenProvider.java
@@ -1,58 +1,0 @@
-package com.vrr.libs.auth.token;
-
-import com.vrr.libs.auth.exception.TokenNotValidException;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.security.Keys;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-
-import java.security.Key;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Date;
-import java.util.stream.Collectors;
-
-@Slf4j
-public class AuthTokenProvider {
-
-    private final Key key;
-    private static final String AUTHORITIES_KEY = "role";
-
-    public AuthTokenProvider(String secret) {
-        this.key = Keys.hmacShaKeyFor(secret.getBytes());
-    }
-
-    public AuthToken createAuthToken(String id, Date expiry) {
-        return new AuthToken(id, expiry, key);
-    }
-
-    public AuthToken createAuthToken(String id, String role, Date expiry) {
-        return new AuthToken(id, role, expiry, key);
-    }
-
-    public AuthToken convertAuthToken(String token) {
-        return new AuthToken(token, key);
-    }
-
-    public UsernamePasswordAuthenticationToken getAuthentication(AuthToken authToken) {
-
-        if(authToken.validate()) {
-
-            Claims claims = authToken.getTokenClaims();
-            Collection<? extends GrantedAuthority> authorities =
-                    Arrays.stream(new String[]{claims.get(AUTHORITIES_KEY).toString()})
-                            .map(SimpleGrantedAuthority::new)
-                            .collect(Collectors.toList());
-
-            log.debug("claims subject := [{}]", claims.getSubject());
-            User principal = new User(claims.getSubject(), "", authorities);
-
-            return new UsernamePasswordAuthenticationToken(principal, authToken, authorities);
-        } else {
-            throw new TokenNotValidException();
-        }
-    }
-}
