--- application/api/src/main/java/com/vrr/libs/config/JwtConfig.java
@@ -1,18 +1,0 @@
-package com.vrr.libs.config;
-
-import com.vrr.libs.auth.token.AuthTokenProvider;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class JwtConfig {
-
-    @Value("${jwt.secret}")
-    private String secret;
-
-    @Bean
-    public AuthTokenProvider jwtProvider() {
-        return new AuthTokenProvider(secret);
-    }
-}
