--- application/api/src/main/java/com/vrr/libs/config/SecurityBean.java
@@ -1,26 +1,0 @@
-package com.vrr.libs.config;
-
-import com.vrr.libs.auth.filter.TokenAuthenticationFilter;
-import com.vrr.libs.auth.token.AuthTokenProvider;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-@RequiredArgsConstructor
-public class SecurityBean {
-
-    private final AuthTokenProvider tokenProvider;
-
-    @Bean
-    public TokenAuthenticationFilter tokenAuthenticationFilter() {
-        return new TokenAuthenticationFilter(tokenProvider);
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-}
